fix(ci): harden deepsec fork scans
